backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.25.0
google-generativeai>=0.8.0
